Upgrade to metalcloud-sdk-go v7.4
Adapt the CLI to the v7.4 SDK, which switched resource IDs and many model
fields from float32/int32 to int64 and reworked several APIs.

int64 migration:
- Use int64/PtrInt64/[]int64/NullableInt64 where SDK models now require it
  (storage, subnet, network_device, logical_network(_profile), endpoint,
  infrastructure, os_template, user, vm_instance, fabric, server_instance,
  bucket, drive, file_share, server_default_credentials).
- Cast int64 model IDs to float32 where API path params remain float32.
- Add utils.GetInt64FromString; fix extension.Revision formatting (%d).

API changes:
- Events: FilterSeverity -> FilterLevel.
- network_device: EnableNetworkDeviceSyslog now returns (job, res, err).
- firmware_catalog: vendor/updateType/severity are plain strings now;
  add firmware_catalog/constants.go with the allowed values.

Removed (no SDK equivalent):
- user limits update, user permissions get/update, infrastructure
  user-limits, and the firmware-policy command group.

Reworked:
- user limits get -> UsersAPI.GetQuotaLimitsBreakdown.
- Startup command-gating is now permissive (the SDK no longer exposes a
  per-user permission list); the server still enforces authorization.

New command groups (SecurityAPI):
- quota-profile (CRUD) and permission create/delete/config-example.

+++ cmd/metalcloud-cli/cmd/endpoint.go
@@ -121,7 +121,7 @@ Examples:
 				}
 			} else {
 				endpointConfig = sdk.CreateEndpoint{
-					SiteId: int32(endpointFlags.siteId),
+					SiteId: int64(endpointFlags.siteId),
 					Name:   endpointFlags.name,
 					Label:  endpointFlags.label,
 				}
